requirements.txt
textual
fastapi
uvicorn
sqlalchemy
pydantic
pyyaml
psutil
alembic
httpx
websockets
